feat(split-webpack): splitting webpack configuration
