+++ webpack-module.config.js
@@ -1,24 +1,16 @@
 const {resolve} = require('path');
-const webpack = require('webpack');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
-
-const pkg = require('./package.json');
 
 /**
  * === ENV configuration
  */
-const ENV = process.env.NODE_ENV;
 const outputPath = resolve('dist');
-const processEnv = {
-  NODE_ENV: JSON.stringify(ENV),
-  appVersion: JSON.stringify(pkg.version)
-};
+const isDev = process.argv.find(arg => arg.includes('webpack-dev-server'));
 
 /**
  * Plugin configuration
  */
-const plugins = [
-  new webpack.DefinePlugin({'process.env': processEnv}),
+const plugins = isDev ? [] : [
   new CleanWebpackPlugin([outputPath], {verbose: true})
 ];
 
@@ -26,12 +18,6 @@ const plugins = [
  * === Webpack configuration
  */
 module.exports = {
-  entry: './src/index.js',
-  output: {
-    path: outputPath,
-    filename: 'module.bundle.js'
-  },
-  devtool: 'cheap-module-source-map',
   module: {
     rules: [
       {
@@ -50,14 +36,6 @@ module.exports = {
             ]]
           }
         }
-      },
-      {
-        test: /\.html$/,
-        use: ['text-loader']
-      },
-      {
-        test: /\.postcss$/,
-        use: ['text-loader', 'postcss-loader']
       }
     ]
   },
